+++ lib/properties.ml
@@ -658,6 +658,7 @@ let pp_property : type a. a property Pp.t =
   | Numeric -> Pp.string ctx "font-variant-numeric"
   | Font_variant_position -> Pp.string ctx "font-variant-position"
   | Font_variant_alternates -> Pp.string ctx "font-variant-alternates"
+  | Font_variant -> Pp.string ctx "font-variant"
   | East_asian -> Pp.string ctx "font-variant-east-asian"
   | Backdrop_filter -> Pp.string ctx "backdrop-filter"
   | Webkit_backdrop_filter -> Pp.string ctx "-webkit-backdrop-filter"
@@ -886,6 +887,7 @@ let property_class : type a. a property -> a property_class = function
   | White_space -> Inherited
   | White_space_collapse -> Inherited
   | Font_variant_alternates -> Inherited
+  | Font_variant -> Inherited
   | Tab_size -> Inherited
   | Webkit_text_size_adjust -> Inherited
   | Font_feature_settings -> Inherited
@@ -1324,6 +1326,7 @@ let property_tag : type a. a property -> int = function
   | White_space -> 199
   | White_space_collapse -> 541
   | Font_variant_alternates -> 548
+  | Font_variant -> 549
   | Border -> 200
   | Border_block -> 201
   | Border_block_start -> 202
@@ -1899,6 +1902,7 @@ let eq_property : type a b. a property -> b property -> (a, b) Type.eq option =
   | White_space, White_space -> Some Equal
   | White_space_collapse, White_space_collapse -> Some Equal
   | Font_variant_alternates, Font_variant_alternates -> Some Equal
+  | Font_variant, Font_variant -> Some Equal
   | Border, Border -> Some Equal
   | Border_block, Border_block -> Some Equal
   | Border_block_start, Border_block_start -> Some Equal
@@ -3211,6 +3215,7 @@ let read_any_property t =
   | "white-space" -> Prop White_space
   | "white-space-collapse" -> Prop White_space_collapse
   | "font-variant-alternates" -> Prop Font_variant_alternates
+  | "font-variant" -> Prop Font_variant
   | "will-change" -> Prop Will_change
   | "word-break" -> Prop Word_break
   | "word-spacing" -> Prop Word_spacing
@@ -3919,6 +3924,7 @@ let canonical_initial_for_minify : type a. a property -> a -> a =
   | White_space, value -> value
   | White_space_collapse, value -> value
   | Font_variant_alternates, value -> value
+  | Font_variant, value -> value
   | ( ( Border | Border_block | Border_block_start | Border_block_end
       | Border_inline | Border_inline_start | Border_inline_end | Column_rule
       | Border_top | Border_right | Border_bottom | Border_left ),
@@ -4879,6 +4885,7 @@ let pp_property_value : type a. (a property * a) Pp.t =
   | White_space -> pp pp_white_space
   | White_space_collapse -> pp pp_white_space_collapse
   | Font_variant_alternates -> pp pp_font_variant_alternates
+  | Font_variant -> pp pp_font_variant
   | Grid_template_columns -> pp pp_grid_template
   | Grid_template_rows -> pp pp_grid_template
   | Grid_template_areas -> pp pp_grid_template_areas